+++ module-audio/Audio/Operation/RecorderOperation.hpp
@@ -35,7 +35,7 @@ namespace audio
 
         Position GetPosition() override final;
 
-        void SetData(std::shared_ptr<void> data) override final;
+        void SetBluetoothStreamData(BluetoothStreamData data) override final;
 
         uint32_t GetSize()
         {