+++ module-services/service-bluetooth/ServiceBluetoothCommon.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <FreeRTOS.h>
+#include <queue.h>
+
+class BluetoothStreamData
+{
+  public:
+    QueueHandle_t in  = nullptr;
+    QueueHandle_t out = nullptr;
+};