+++ changelog.md
@@ -7,6 +7,7 @@
 * `[meditation]` Added gong sound on meditation intervals
 * `[meditation]` Added basic meditation functionalities and settings
 * `[cellular]` Handling incoming calls and sms in sleep mode
+* `[audio]` Added Bluetooth related APIs to AudioServiceAPI
 
 ### Fixed